- Add annotations to OCLogger, so that Xcode now checks format strings and issues warnings for wrong or missing parameters - Fixed all issues surfaced by these warnings

+++ ownCloudSDK/Vaults/Database/SQLite/Internals/OCSQLiteDB+Internal.m
@@ -34,7 +34,7 @@
 	{
 		if (_liveStatements.count > 0)
 		{
-			OCLogDebug(@"Releasing the resources of up to %d live statements", _liveStatements.count);
+			OCLogDebug(@"Releasing the resources of up to %lu live statements", (unsigned long)_liveStatements.count);
 		}
 
 		for (OCSQLiteStatement *statement in _liveStatements)
